Add files via upload
This is one of the first data cleaning jobs I've done, was done with guidance from skills network. The aim of this project was to understand the steps of data cleaning, what actions can be taken with numpy and pandas to ensure a dataframe is ready for EDA. This got me some good progress with knowing how to ensure data is valid and useful for later steps in data science. Handling erroneous or spurious includes knowing the expected range of a column's data which can be gleaned from the metadata and or documentation of the original data.